Move fee tests to new module
