+++ crates/astria-sequencer/src/app/mod.rs
@@ -3,6 +3,8 @@ pub(crate) mod test_utils;
 #[cfg(test)]
 mod tests_app;
 #[cfg(test)]
+mod tests_block_fees;
+#[cfg(test)]
 mod tests_breaking_changes;
 #[cfg(test)]
 mod tests_execute_transaction;
