+++ crates/astria-sequencer/src/bridge/init_bridge_account_action.rs
@@ -20,7 +20,7 @@ use crate::{
     app::ActionHandler,
     assets::{
         StateReadExt as _,
-        StateWriteExt,
+        StateWriteExt as _,
     },
     bridge::state_ext::{
         StateReadExt as _,
@@ -104,7 +104,8 @@ impl ActionHandler for InitBridgeAccountAction {
         );
         state
             .get_and_increase_block_fees(&self.fee_asset, fee, Self::full_name())
-            .await?;
+            .await
+            .context("failed to get and increase block fees")?;
         state
             .decrease_balance(from, &self.fee_asset, fee)
             .await
